pyproject.toml
[project]
name = "ai-news-aggregator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0.12",
    "youtube-transcript-api>=1.2.4",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.48",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "google-genai>=0.3.0",
    "html-to-markdown>=3.1.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
]
